chore(studio): move DocumentationLink next to ThemeSwitch
Route-agnostic top-bar widgets live in routes/PageLayout (see
ThemeSwitch), while GlobalNav itself only hosts the shell. Moves the
new documentation link there to match.
