+++ web/packages/studio/src/components/Layouts/GlobalNav/index.tsx
@@ -3,12 +3,12 @@
 
 import { AppBar, Button, Flex, Stack, Text, Tooltip } from '@nvidia/foundations-react-core';
 import { Breadcrumbs } from '@studio/components/Breadcrumbs';
-import { DocumentationLink } from '@studio/components/Layouts/GlobalNav/DocumentationLink';
 import { UserPopover } from '@studio/components/UserPopover';
 import { TOUR_ENABLED } from '@studio/constants/environment';
 import { ROUTES } from '@studio/constants/routes';
 import { useWorkspaceFromPathIfExists } from '@studio/hooks/useWorkspaceFromPath';
 import { AssistantTopBarChat } from '@studio/routes/agents/AssistantChatRoute/AssistantTopBarChat';
+import { DocumentationLink } from '@studio/routes/PageLayout/DocumentationLink';
 import { ThemeSwitch } from '@studio/routes/PageLayout/ThemeSwitch';
 import { getWorkspaceDetailsDefaultRoute } from '@studio/routes/utils';
 import { useSidebarState } from '@studio/util/hooks/useSidebarState';
