+++ web/packages/studio/src/routes/PageLayout/DocumentationLink.test.tsx
@@ -1,8 +1,8 @@
 // SPDX-FileCopyrightText: Copyright (c) 2025-2026 NVIDIA CORPORATION & AFFILIATES. All rights reserved.
 // SPDX-License-Identifier: Apache-2.0
 
-import { DocumentationLink } from '@studio/components/Layouts/GlobalNav/DocumentationLink';
 import { LINK_DOCS_STUDIO } from '@studio/constants/links';
+import { DocumentationLink } from '@studio/routes/PageLayout/DocumentationLink';
 import { render, screen } from '@studio/tests/util/render';
 
 describe('DocumentationLink', () => {
